+++ ricaun.DA4R.NUnit/Extensions/ZipExtension.cs
@@ -23,7 +23,7 @@ namespace ricaun.DA4R.NUnit.Extensions
         {
             var zipName = Path.GetFileNameWithoutExtension(zipFile);
 #if DEBUG
-            zipName = $"Temp_{DateTime.Now.Ticks}";
+            zipName = $"{zipName}_{DateTime.Now.Ticks}";
 #endif
             var zipDirectory = Path.GetDirectoryName(zipFile);
             zipDestination = Path.Combine(zipDirectory, zipName);
@@ -50,6 +50,36 @@ namespace ricaun.DA4R.NUnit.Extensions
             catch { }
 
             return false;
+        }
+
+        /// <summary>
+        /// ZipCurrentFolder
+        /// </summary>
+        /// <param name="zipFileName"></param>
+        /// <param name="excludedFiles"></param>
+        public static void ZipCurrentFolder(string zipFileName, params string[] excludedFiles)
+        {
+            string sourceFolderPath = Directory.GetCurrentDirectory();
+            string destinationZipPath = Path.Combine(sourceFolderPath, zipFileName);
+
+            // Create a new ZIP file
+            ZipFile.CreateFromDirectory(sourceFolderPath, destinationZipPath, CompressionLevel.Optimal, false);
+
+            // Open the created ZIP file
+            using (var zip = ZipFile.Open(destinationZipPath, ZipArchiveMode.Update))
+            {
+                // Iterate over the excluded files
+                foreach (var excludedFile in excludedFiles)
+                {
+                    // Get the entry for the excluded file if it exists
+                    var entry = zip.GetEntry(excludedFile);
+                    if (entry != null)
+                    {
+                        // Remove the entry from the ZIP file
+                        entry.Delete();
+                    }
+                }
+            }
         }
     }
 }