+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] / 2023-05-30
+### Local
+- Add `AppDomainUtils` to force Load `ricaun.NUnit` dependencies.
+- Add `SampleTest.zip` file in `Debug`
+### Updated
+- Update `TestService`
+- Update `ricaun.NUnit` to version `1.2.9`
+
 ## [1.0.1] / 2022-12-13
 ### Updated
 - Update ricaun.NUnit to version `1.1.0`
@@ -20,4 +28,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - TestAssembly in `*.zip` and `output.json`
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.2]: ../../compare/1.0.1...1.0.2
+[1.0.1]: ../../compare/1.0.0...1.0.1
 [1.0.0]: ../../compare/1.0.0